feat(ai): allow managers to use planning for their department only

+++ backend/src/main/java/com/capevents/backend/dto/ai/planning/AiPlanningEventProposalRequest.java
@@ -8,7 +8,7 @@ public record AiPlanningEventProposalRequest(
         String referenceDate,
 
         @JsonProperty("target_department_id")
-        Integer targetDepartmentId,
+        Long targetDepartmentId,
 
         Integer limit,
